[general] added TGetWindowText() function which gets the window text as a TString. made some changes to mIRC_DrawText() to fix the multiline text problems. NB: NOT PERFECT YET.
[mdialog]
changed to set back to OrigID in destroy, instead of setting to ID.

[toolbar]
removed some commented code.
changed to use MapWindowPoints() instead of ClientToScreen() & ScreenToClient()
noticed bug in tooltip contents, they will always be the contents of the button you hover over first. (dunno why, working on it)

[box]
changed to use TGetWindowText()

[text]
removed the transparent style handling as this is done globally now.
changed to use TGetWindowText()
re-applied the changes made before to allow the the formatted text to be displayed when in non-transparent mode, SOMEONE must have removed my changes, wonder if anything else was removed.