[general] changed /xdid -J to use LoadImage(IMAGE_CURSOR) & get a non-shared cursor handle, this allows us to delete the cursor & fixes the graphics corruption bug. changed all cursor code as needed to delete the LoadImage() cursor. changed xdid -J to check file exists added an error report to xdid -J when the cursor fails to load. changed BitmapRegion() to show an error when CreateDIBSection() fails. fixed a small mem leak in BitmapRegion() when CreateDIBSection() fails.
[dialog]
changed /xdialog -q to use LoadImage(IMAGE_CURSOR) & get a non-shared cursor handle, this allows us to delete the cursor & fixes the graphics corruption bug.
changed all cursor code as needed to delete the LoadImage() cursor.
changed /xdialog -q to check file exists
added an error report to /xdialog -q when the cursor fails to load.

[divider][button][calander][colorcombo][comboex][datetime][directshow][listview][mdialog][mwindow][pager][panel][pbar][rebar][richedit][stacker][statusbar][tab][toolbar][trackbar][treeview][updown][box][check][edit][image][line][link][list][radio][scroll][text]
changed WM_SETCURSOR code to check if the cursor to set is the same as the current before trying to set.

[pager]
changed xdid -d to remove a mdialog or mwindow control from the internal list.