[general] added showError() function to replace all other error display functions.   atm this function will echo the error or send to dcx signal if active, this is likely to change.
[dialog]
added showError() function to replace all other error display functions.
  atm this function will echo the error or send to dcx signal if active, this is likely to change.
changed xdialog -j to only redraw the whole dialog if an ID wasn't given.

[directshow]
added $xdid().isloaded property to tell if a file is loaded or not.
changed dcxInfoError() to be showError() instead.

[image]
changed dcxInfoError() to be showError() instead.
re-enabled the background erase code, fixes image not updating bug.
added commented code for a different way of fixing that to xdid -a, if needed.

[list]
changed dcxInfoError() to be showError() instead.