[dll] changed the $debug check  in LoadDll() to use the new TString constructor. changed the xpopup class setup to use the already defined WNDCLASSEX changed the xpopup owner window to be of type HWND_MESSAGE on XP+
[toolbar]
fixed /xdid -q bug where the command id was used but the TBIF_BYINDEX flag was specified.
changed /xdid -c & -i to use the TBIF_BYINDEX flag.

[xtreebar]
added check for valid - char in switch.
improved the handling of a non-winsdk compile.
added an enum for all the /xtreebar -s styles.
added /xtreebar -s balloon & noballoon styles to change the tooltips into balloons or not.
 This can only be done after tooltips have been enabled. (NOT IN THE SAME /xtreebar -s THAT ENABLED THE TOOLTIPS)
changed invalid style error to return the style in question.
changed /xtreebar -c to check for a valid + char in flags.
changed the image list we create in /xtreebar -w to use the icon sizes in the default list instead of 16x16 (this is still the default if the sizes can't be found)