[general] added ctrlDrawText() to DcxControl which draws the text depending on the controls shadow etc.. settings. changed some mIRCDebug() calls to DCXError() fixed memleak bug in DCXErrorEX() caused by missing va_end() call.
[dialog][tab]
changed a mIRCDebug() to be a showError()

[tab][box][line][list][text]
changed to use ctrlDrawText()

[link]
changed to use the xdid -f set font when it exists, instead of always the default gui font.

[list]
changed wm_drawitem to propperly set bParsed.
changed to allow setting the normal text colour via xdid -C +t

[xpopup]
some misc formatting changes.
changed DrawItemText() to check if the first char of the text is $chr(12) & enable utf8 support for that item if it is. (only way i could thing of to allow per-item utf8 atm)