[general] changed some vars to consts. removed some non-unicode code.
[dcxml]
fixed listview icon loading bug which caused all icons to be added to state list.

[xtreebar]
changed geticons callback behaviour, now sets I_IMAGENONE when <1 is returned for the icon index instead of zero.

[xstatusbar]
changed status_parseControlStyles()/status_parseItemFlags()/status_cleanPartIcons() to use a for() loop.
changed /xstatusbar -l to use a for loop.
changed /xstatusbar -w to move the status_getImageList() call out of the if().
changed $xstatusbar().parts to use a for() loops.

[xdock]
changed DockWindow() to use XSwitchFlags.